corrigido os desafios

+++ Desafio007.py
@@ -2,8 +2,8 @@ print('====== DESAFIO 007 ======')
 
 # Densenvolva um programa que leia as duas notas de um aluno, calcule e mostre a sua média.
 
-n1 = int(input('Informe a primeira nota: '))
-n2 = int(input('Informe a segunda nota: '))
+n1 = float(input('Informe a primeira nota: '))
+n2 = float(input('Informe a segunda nota: '))
 
 media = (n1 + n2)/2
 print('A média do aluno é {}'.format(media))
