+++ Desafio008.py
@@ -2,7 +2,7 @@ print('====== DESAFIO 008 ======')
 
 # Escreva um programa que leia um valor em metros e o exiba convertido em centimetros e milímetros.
 
-metros = int(input('Informe a quantidade de metros: '))
+metros = float(input('Informe a quantidade de metros: '))
 
 centimetros = metros * 100
 milimetros = metros * 1000
