+++ Desafio012.py
@@ -2,7 +2,7 @@ print('====== DESAFIO 012 ======')
 
 # Faça um programa que leia o preço de um produto e mostre seu novo preço, com 5% de desconto.
 
-produto = int(input('Informe o preço do Produto: '))
+produto = float(input('Informe o preço do Produto: '))
 
 desconto = produto * 0.05
 novo_preco = produto - desconto
